Fix bug where data JSON file is not included in the build
Bug description: previously, the `menu_items.json` file, which contains
all the initial menu items, was mistakenly left out of the build i.e.
webpack didn't copy it to the build directory. If we take this build
directory outside the project directory, the app will fail to load data
on launch. Up to this point, the application only fetched data
successfully only because the build directory is at the same level as
the `src` directory.

This is now fixed by using `file-loader`.